run_agent.py
import sys
import os
import openai
import yaml

def load_knowledge_hub(task):
    # 根據任務加載相關的知識庫文件
    # 返回格式化的字符串
    pass

def generate_ai_response(prompt):
    # 使用 OpenAI API 生成回應
    pass

def determine_output_file(task):
    # 根據任務名稱生成輸出文件名
    return f"{task}_output.txt"

def main(task):
    # 1. 加載任務相關的知識
    knowledge = load_knowledge_hub(task)
    
    # 2. 構建提示詞
    prompt = f"""
    Task: {task}
    Knowledge Base:
    {knowledge}
    
    Please generate the required output based on this information.
    """
    
    # 3. 獲取 AI 響應
    response = generate_ai_response(prompt)
    
    # 4. 將響應寫入適當的文件
    output_file = determine_output_file(task)
    with open(output_file, 'w') as f:
        f.write(response)

if __name__ == "__main__":
    if len(sys.argv) > 1:
        task = sys.argv[1]
        main(task)
    else:
        print("Usage: python run_agent.py <task>")
        sys.exit(1)